Move GridInfo to its own module
